src/requirements.txt
py-solc==2.1.0
web3==3.16.4
Flask==0.12.2